[change-owners] empty self-service must render a change non-self-serviceable
If a change is covered by a self-service role without members, nobody can approve the change while at the same time the MR is still marked as self-serviceable. This way it does not show up in the review queue.

But even if IC gets aware of the MR, they can't /lgtm or add an lgtm label. The only way out is a manual force merge by IC.

Refine the term self-serviceable. If a change is only covered by a self-service role without members, the change is not considered self-serviceable. This way AppSRE IC will get aware of the change via the review queue and can approve the regular way.

The change coverage report in the MR should still show that the change would be self-serviceable but is not because of the empty role.

part of APPSRE-7535

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>