[change-owners] refactor metadata handling (#3501)
- use pydantic dataclasses to parse and validate data from the qontract-server `/diff` endpoint
- split content data and metadata from file changes and refactor metadata change detection

the main point of this PR is the handling of metadata changes, e.g. if a file gets moved or if just a SHA changes without any content changes (e.g. YAML comments not being present in the qs bundle). previously metadata and data were mixed in `BundleFileChange` objects. now metadata fields are wiped from the actual file contents and are stored seperately. this simplified a lot of code revolving around self-service for metadata changes.

because of this change, metadata changes like the path of a file are now self-serviceable.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>