[cluster-upgrades] use available upgrade versions data from cluster spec
the `OCM.get_available_upgrades()` function gets available upgrades for a specific version in a specific channel. behind the scenes it uses the CS `/v1/versions` endpoint, which potentially fails requests for non stable channels if the requesting account/service-account belongs to an OCM organization that lacks the capability to use such versions.

using this endpoint with a service-account that is not bound to an OCM organization, always fails.

this PR changes the datasource for the available upgrades information to the`version` section of the OCM cluster spec, which can be safely consumed by service accounts.

`OCM.get_available_upgrades()` is changed to accept a cluster name instead of a current version and channel.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>