Limit the size of Tekton Pipeline names
Pipeline names are built using the following convention:

```
o-<one per saas file pipeline name>-<saas file name>
```

The max size of `<saas file name>` was chosen to be 39 characters taking
into account that pipeline name was `openshift-saas-file`, so that we
can have pipelines with names up to 63 characters. This also worked well
with our PipelineRuns, which appended a 14 characeters date but trimmed
the Pipeline size so that, again, objects were right in size.

This didn't take into account that Tekton retries create PipelineRun names
that are named appending a 7 character random string to the corresponing
Pipeline name, which make certain Pipeline names not possible.

This patch makes sure that we have Pipeline names that will be suitable
for Tekton retries. It also changes many fixtures that use the old and
not longer allowed long pipeline name so that they are more realistic.

The code protects beyond what the schema will allow, since pipeline
names are also being limited in their length.

part of APPSRE-6031

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>