[openshift-resources] enforce managed-resources names (#3589)
it seems that we lost the ability to enforce managedResourceNames for openshift resource in namespaces (or it never was enforced, who knows).

this PR manages the allowed resources names as part of the `ResourceInventory`. the allowed names are registered along with the allowed resource types during clusteR/namespace/type registration. if a desired resources is added to the inventory with a name that is not managed, an error is raised. the resource inventory will be marked as failed, so the execution will fail (important for PR checks).

the managedResourcesNames remain optional. if none are defined, the resource inventory acts as before, allowing all names.

i'm unsure if we should also fail adding current state resources to the inventory if their name is not managed. i believe it is not necessary right now because if managed resource names are defined, the fetching processes for a namespace/kind takes care of filtering already.

https://issues.redhat.com/browse/APPSRE-7811

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>