Add alerts report from sd-app-sre-alert channel
Create a report parsing the output of the conversations.list API
endpoint for our #sd-app-sre-alert channel that can be filtered by days
from the moment that report is run.

We also have explicit timestamp to be able to recreate a certain report
from the past.

This is a `get` qontract-cli command as the rest of the similar reports
we have.

It generates a table with the following colums:

* Alert name
* Triggered
* Resolved
* Median time to resolve

A few comments about the potential scenarios:

* Triggered doesn't need to be equal to resolved, as triggering message
  can be outside the considered time window.
* Median time to resolve sometimes cannot be determined as we use alert
  messages for that that can contain dynamic parts.

part of APPSRE-7762

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>