[AUS] reference version history of an org around ocm-env + org-id (#3613)
change the way the version history of an OCM organization is referenced. right now it is referenced by OCM organization name. this is problematic when

1) the name of an organization changes (rarely)
2) the name of an organization is not unique

especially organization name duplicates are problem in environments where the name is not controlled by us, e.g. AMS organization names.

please note that this PR does not contain any migration logic for the state. state keys need to be renamed manually before these integrations run.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>