[AUS] frozen condition for cluster upgrade policies
Sometimes an upgrade for a cluster needs to prevented. The current ways of setting high soakdays or removing the policy are not nice because for both the fact that the cluster is just temporary frozen is not apparent.

Adding a new `frozen` condition to upgrade policies allows to define explicit upgrade freeze.

part of https://issues.redhat.com/browse/APPSRE-7870

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>