[change-owners] relate change-types to all files of a schema
to further drive self-service we need to be able to relate a change-type to all files of a certain schema. this is useful if we want to define admin roles, e.g. self-service-admins can approve all self-service related changes to roles or change-type admins can approve all changes to change-types without the need to relate a change-type to all datafiles/resources.

```yaml
self_service:
- change_type:
    $ref: /app-interface/changetype/change-type-admin.yml
- change_type:
    $ref: /app-interface/changetype/self-service-admin.yml
```

In this example the change-types are mentioned without listing datafiles or resources, hence the declaring role can approve whatever the change-type defines on all files that match the change-types `contextSchema`.

https://issues.redhat.com/browse/APPSRE-7876

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>