[rhidp] fix multiple IDPs with same username error (#3620)
This commit introduces a bugfix for the "Could not create user"cluster login error (APPSRE-7869) for users whose GitHub and the RedHat usernames are the same.

This commit changes the mapping_method OIDC configuration from claim to add.

This will change all new and existing OIDC IDPs but only for new GitHub IDPs (no update of the existing config).

claim
The default value. Provisions a user with the identity’s preferred user name. Fails if a user with that user name is already mapped to another identity.

add
Provisions a user with the identity’s preferred user name. If a user with that user name > already exists, the identity is mapped to the existing user, adding to any existing identity mappings for the user. Required when multiple identity providers are configured that identify the same set of users and map to the same user names.

Ticket: APPSRE-7872