[AUS] refactor all AUS flavors
the goal of this refactor was to concentrate most of the required queries to determine the state into the `get_ocm_env_upgrade_specs` method of the integration class. it was hard to understand what data was loaded where and the same data was loaded multiple times due to that. the heavy use of the `OCM/OCMMap` classes in various places of the code were witness of that.

what happened:

* extracted version-gate, version-gate agreement, policy mgmt functions from the `OCM` class into reconcile.utils.ocm since they are not organization related. this allows to have lightweight context during these operations and simplifies testing
* ClusterUpgradeSpec is used throughout the code as desired state instead of the Configured* classes. this allows us to keep way more context that is not lost during model translation and hence we don't need to query missing data again via `OCM/OCMMap`
* translated all related `unittest` tests to `pytest`
* reuse the `get_ocm_env_upgrade_specs` functionality in `qontract-cli` to get rid of custom state compilation (extensive testing was done here. the resulting reports are still the same as previously to the letter)

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>