[change-owners] make context conditions available without explicit context selector
`jsonPathSelectors` of a change-type can now be used conditionally with the `context.when` condition even when no `context.selector` is specified. a missing context selector defaults to `$.path` which selects the file where the actual change happened. this was not possible before because of a bug.

this enables change-types that react to file removals/additions directly. previously this was only available indirectly.

e.g. the following change-types grants permissions to self-service the removal

```yaml
---
$schema: /app-interface/change-type-1.yml

name: jira-board-auto-remover

priority: low

contextType: datafile
contextSchema: /dependencies/jira-board-1.yml

changes:
- provider: jsonPath
  jsonPathSelectors:
  - $
  context:
    [selector: $.path]  <-- optional because it defaults to this
    when: removed
```

https://issues.redhat.com/browse/APPSRE-7940

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>