[ocmmap] basic cluster filtering during CS request
filter for `managed` clusters and filtering on the `product` (OSD/ROSA/Hypershift) is now done as part of the CS request instead of filtering the raw result afterwards. this reduces the number of queries a lot in big organizations with clusters that dod not fit those filters.

filtering for state `ready` is still on the raw result because certain code relies on knowing about non-ready clusters. we can reconsider this in a later PR.

part of https://issues.redhat.com/browse/APPSRE-8152

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>