Capture errors from secrets in saasherder
We currently silently fail when saasherder cannot populate desired
state due to secret management issues.

This patch also contain a safeguard to make sure that populate desired
state does log something (in DEBUG, to avoid duplicates in prod) so that
debugging is feasible. Not the cleanest, but a better solution would
require a whole refactoring of _process_template.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>