[dynatrace-token-provider] New integration to distribute dynatrace access tokens to Hypershift clusters (#3734)
The integration should be able to retrieve clusters with "sre-capabilies.dtp*" labels, filter them against an allow list of OCM org IDs passed through parameters, then create two tokens with specific scope in Dynatrace for each cluster and make sure that those tokens are distributed to the cluster through SyncSet. It should also handle the situation where the access token was accidentally deleted from Dynatrace side, under which situation it will re-create the missing token and patch the SyncSet. It assumes the integrity of SyncSet once created, since nothing else should be altering this specific SyncSet. 

To support this work, an ocm syncsets module was added. dynatrace as a dependency was introduced. As a result of it not being typed, it was added to exception list.