[sharded dry-runs] speed up shard detection (#3784)
* [sharded dry-runs] speed up shard detection

shard detection turned out to be slow on bigger desired states like on prometheus-rules tester and openshift-resources.

in this MR the approach has been simplified for these integrations by returning a desired state that encodes the shards as keys in the state and uses a hash for all the configs that belong to that shard.

this makes the desired state to operate on consistently small and all actions are way faster. for hash calculation, deepdiff.Deephash is used, which produces consistent hashes in a fast way.

before this change, prometheus rules tester took > 15s to find the affected shards, now it does the same in ~2s.

also, i believe we should restructure the early exit desired state of all integrations to have their shards encoded as top level keys. this way, shard detection could be achieved without defining jsonpaths to the places where shards can be found in the desired state because it is found in the same place for all integrations. harmonization ftw.