expose cluster upgrade state via metrics
we lack upcoming and ongoing upgrade information on the AUS dashboards.

this patch encodes such data as negative values in the `aus_cluster_version_remaining_soak_days` metric.
this is safe since the metric currently has a min value of 0.

we will use -1 for scheduled upgrades (upgrade policy created in OCM) and -2 for started/ongoing upgrades.

part of https://issues.redhat.com/browse/APPSRE-7878

additionally the channel group of a cluster is exposed in the AUS info metric for dashboarding purposes.

part of https://issues.redhat.com/browse/APPSRE-7877

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>