[AUS] highlight blocked versions on dashboard
if a version is blocked, the metric for remaining soak days will expose -4 instead.

additionally the remaining soak days metric will not be exposed for versions if another followup (non-blocked reachable version) with 0 soak days is already available. this way we don't clutter the dashboard with version columns that bear no real value because such versions will not be considered for upgrade. important events like currently scheduled or running upgrades, or blocks will still be exposed.

APPSRE-7878

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>