Add `eval` filter to jinja2
We have a use case where we want to render the slo documents into
PrometheusAlerts.

The fields in the slo-documents datafile do contain Jinja2 parameters,
for example:

```
expr: |
    sum(increase(aggregation:quay_request_duration_seconds_count:rate1m:sum{route="api.repository",status!~"5[0-9][0-9]",method="GET"}[{{window}}]))
    /
    sum(increase(aggregation:quay_request_duration_seconds_count:rate1m:sum{route="api.repository",method="GET"}[{{window}}]))
```

The `{{window}}` parameter must be evaluated inside the jinja2
expression:

```
{{% set expr = slo.expr | eval(window=slo.SLOParameters.window) | trim + ' <= ' + slo.SLOTarget | string %}}
```

In order for the `eval` part to work, this filter is required.