use a-i state via STS (and more ways to connect to a-i state) (#3827)
until now, state initialization was suppored based on
* AWS account stored in app-interface
* state in app-interface-settings

with this MR, the following ways are supported

* env vars APP_INTERFACE_STATE_BUCKET, APP_INTERFACE_STATE_BUCKET_REGION, APP_INTERFACE_STATE_AWS_PROFILE and an AWS config or creds file is available
* env vars APP_INTERFACE_STATE_BUCKET, APP_INTERFACE_STATE_BUCKET_REGION, APP_INTERFACE_STATE_VAULT_SECRET (and optionally APP_INTERFACE_STATE_VAULT_SECRET_VERSION)
* env vars APP_INTERFACE_STATE_BUCKET, APP_INTERFACE_STATE_BUCKET_REGION, APP_INTERFACE_STATE_BUCKET_ACCESS_KEY_ID, APP_INTERFACE_STATE_BUCKET_SECRET_ACCESS_KEY
* env vars APP_INTERFACE_STATE_BUCKET, APP_INTERFACE_STATE_BUCKET_REGION and APP_INTERFACE_STATE_BUCKET_ACCOUNT if the mentioned AWS account is present in app-interface
* state settings in app-interface-settings-1.yml

especially the mechanism via `APP_INTERFACE_STATE_AWS_PROFILE` enables us to leverage STS for state access.

important: the new way to initialize state is fully compatible to the previous way. all existing env variables are honored and the new mounts and secret based env vars are set to optional so that existing q-r deployments still work without the need for change.