Ensure the same SA token is always copied (#3908)
This patch eliminates the condition that checks if a token is already
present in the destination namespace so that it is that one that is
copied instead of the first of the list. This made that in case we had
multiple tokens associated to an account, we could potentially copy to
different tokens to two different clusters if there was already a token
copied that it was not the first of the `sa_token_list` list. This also
made the Vault shared-resources secret to be updated multiple times in
each run if we had different tokens being copied.

fixes APPSRE-8480

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>