pluggable membership sources for roles (#3909)
We should be able to run integrations that rely on roles, users and role membership in the sre-capabilities repo with as little change as possible, while not being forced to manage a whole set of users and role memberships.

The central concept for grouping and permissions will still be `/access/role-1.yml` but membership will be pluggable. A configuration section memberSource can be used to bring in user and role membership information from different sources, defined via a schema `/access/membership-provider-1.yml`.

A membership provider can be any system that provides groups and users (provider pattern). User information needs to include some sort of username that can be mapped to org_username. This schema change introduces the memebership provider mechanism and offers app-interface as a first source for members.

```yaml
$schema: /access/membership-provider-1.yml

name: <provider-name>
source:
  provider: app-interface
  url: <url to qontract server of an a-i instance>
  username:
    $ref: <vault secret>
  password:
    $ref: <vault secret>
```

Then this provider can be used as follows

```yaml
---
$schema: /access/role-1.yml
...
memberSources:
- provider:
    $ref: <the provider>
  group: <a role name from the other app-interface>
```

depends on https://github.com/app-sre/qontract-schemas/pull/546

APPSRE-8450