import OCM cluster OIDC endpoint URL into a-i cluster files (#3954)
Import the OIDC endpoint URL of ROSA clusters into app-interface cluster files, similar to the ID and external ID of a cluster. It is supposed to be read only information that can be used by people and automation to define trust policies in IAM policies without the need to do OCM lookups. Since this URL is static, "caching" it like this in the cluster file is a legit approach.

part of https://issues.redhat.com/browse/APPSRE-8656

depends on https://github.com/app-sre/qontract-schemas/pull/560

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>