[tf-vpc-peerings] allow access to private HCP API VPCE via peering
the HCP API is offered via a VPC endpoint in the clusters AWS account. access to this VPCE is per default only granted from within the VPC. when accessing the API via VPC peering, the security group of the VPCE is preventing communication.

this patch will add new rules to the security group of the VPC endpoint. sadly we can't create a dedicated SG and add it to the VPCE because that is only possible with the AWS TF provider > 4 (and we are at 3.7.x). so our only way is to add more rules to the existing SG.

i checked the code of the hypershift operator that reconciling this SG. it is not removing any excessive rules, only adding missing ones. once we upgrade to a newer TF provider, we can add new security groups to existing VPC endpoints. but for now we need to accept this approach.

part of https://issues.redhat.com/browse/APPSRE-8666

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>