APPSRE-8160: Use the upstream regular expression for validation
This will take the check "further" than what was requested in the Jira card, but will provide more future proofing.
This change also clarifies a code comment and the way the test functions by feeding the method an actual dict (which it would be receiving anyways). As a result, the test table has grown a little in order to accomodate the extra value information alongside the key information.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>