[AUS] expect at least one version gate for a minor version (#3976)
change: if a minor version has no version gates, upgrades to that minor versions are skipped.

reasoning: usually every minor version has at least one version gate in place. this is not technically necessary but practically speaking this has been the case for every minor version of the last couple of years. we want to skip scheduling an upgrade in such situations because missing version gates might be an indicator that the gap analysis has not been conducted yet from the platform team. such upgrades bear many unknowns and are not considered safe. this mostly happens early during pre-GA, e.g. for RCs.

https://issues.redhat.com/browse/APPSRE-8694