Add template validation CLI
This CLI can be used to develop new templates. It eliminates the need to
start qontract server and run the template validation integration. It
additionally enables using yamllint to ensure the created template meets
our lintint specification. Also it can run qontract-validator to ensure
the created template is adhering qontract-schema