native deepdiff 6.7.x path parsing and tf-res sharded pr check fix
the deepdiff 6.7.x library yields different path formats for the found
diffs. our custom parse logic was not able to deal with that but
deepdiff now also offers a native way to parse paths.

this PR uses the deepdiff path parse logic instead of our own.
additionally a bug in the produced early exit desired state was fixed.
the state can only contain dicts and not dataclasses.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>