[job-controller] support secret data for jobs
the `K8sJob` class can define secret data it relies on and the job
controller will manage the lifecycle of a kubernetes Secret based on
that data.

```python

class MyJob(K8sJob):

   ...

   def secret_data(self) -> dict[str: str]:
     return {
       "super": "secret"
     }

```

the resulting secret has an `ownerReference` towards the job, which
implies that the secret gets cleaned up when the job gets cleaned up.

since these kind of jobs leverage `ttlSecondsAfterFinished` to
self-destroy, also the secrets lifecycle depends on the job ones.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>