[rosa] leverage STS for ROSA automation within the AWS account (#4102)
leverage in-cluster STS functionality to interact with the AWS accounts
for ROSA clusters.

this PR also removes the cluster context from the `RosaSession`. the
session should be about the AWS and OCM context, not demanding
information about a cluster. this makes the rosa utilities more
versatile during cluster creation or account dressup for cluster
creation.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>