[version gate approver] creates version gate agreements when conditions are met (#4103)
not all version gates can be agreed to just like that. the STS gates require some actions on the clusters account and even the acking the OCP gate just like that is debatable.

this new integration checks for different version gate conditions and conditionally creates version gate agreements. in this first iteration it takes care the STS gate, by upgrading the ROSA account and operator roles and acknowledges the gate to allow AUS to proceed in upgrading. at the same time the AUS integration is no longer trying to acknowledge version gates but expects all of them to be acked.

upgrading roles requires access to the AWS account hosting the cluster. this integration uses AWS account credential secrets and AssumeRole to these accounts if allowed. this is fully compatible with kube based service account STS if the AWS credentials leverage AssumeRoleWithSTS. the actual role upgrade happens based on the job system defines in reconcile.utils.jobcontroller and the ROSA job utility system from reconcile.utils.rosa.

this integration can run as an AUS sub-capability, meaning it does not require clusters, accounts and access credentials defines in GQL. clusters are detected via OCM labels, OCM authentication is established by OCM organization files in GQL (which is the way to invite an OCM organization to partace in capabilities) and AWS access is defines by respective kube secrets holding jump credentials.

a cluster needs to opt in with a sre-capabilities.aus.version-gate-approvals OCM subscription labels to have their gates selectively managed. for the STS one it means that the owner of the cluster and holder of the AWS accounts requires to grant assume role permissions to this integration. for app-interface managed clusters, this label is managed based on this PR

a lot of things showcased in this PR will be transferable to other usecases, e.g. cluster creation