[AUS] cluster health checks (#4187)
a cluster will only provide soak days for a version if it is healthy. health checks are brought into the AUS context via the `reconcile.utils.clusterhealth` module, which currently leverages the telemeter instance of an OCM environment to look for critical cluster alerts.

health checks sources need to be enabled in the sre-capabilities repo in the OCM org file as follows

```yaml
$schema: /openshift/openshift-cluster-manager-1.yml
...
ausClusterHealthChecks:
- provider: telemeter
  enforced: true/false
```

currently, only `telemeter`is available as a supported provider.

a health check provider can be set to be in enforced mode or not. `enforced` means, that an unhealthy cluster will not provide soak-days. otherwise health issues will only be highlighted in the logs and with metrics.

the cluster health is exposed as a metric `aus_cluster_health_state`. there is one such metric exposed per cluster and health check provider.

depends on https://github.com/app-sre/qontract-schemas/pull/601

part of https://issues.redhat.com/browse/APPSRE-8789