[change-owners] approval detection for split changes
consider a path in a datafile approved if all individual parts of the
datafile are covered by change-types and have been approved.

e.g. consider the introduction of an RDS overrides section in a
namespace

```yaml
externalResources:
- provider: aws
  provisioner:
    $ref: /aws/account.yml
  resources:
  - provider: rds
    ...
+   overrides:
+     timeouts:
+       create: 2h
+      apply_immediately: true

```

the entire `overrides` section is covered by a high priviledge
change-type and the `timeouts` and `apply_immediately` properties are
covered by two other specialized change-types.

this PR will allow the change to be self-serviceable by:
* the high priviledge change-type...
* ... or the two specialized change-types, without the need of the high priviledge change-type

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>