Template mr management (#4220)
* Add basic MergeRequestManager

Add a parser to parse open MergeRequests

* Additional code for merge request manager

* Add GitlabFilePersistence

* Various fixes

* Move things around to make code base easier to read

* Various small fixes

* Adding state handling

By this, templates are not rendered, when they where rendered in the past and have not changed. Variable input is also taken into account for this. A state key is kept for each collection, a has is computed over the template content and the variable content

* Adding tests for Merge Request Manager

* Adding renderer tests

* Adding tests for gitlab persistence

* Fix wrong import

* Test filenames need to be unique https://github.com/pytest-dev/pytest/issues/3151

* Fix commit message, rename template_hash, since it's a hash of the collection

* Slight adjustments

Allow closing MRs
Directly open MRs after closing them