Introduce persistence context
Context holds the value of a file managed by a template collection.

This fixes a bug, where a collection updating a file multiple times using more then one template, could lead to missing updates. Reason was, that previously everytime the file was read from persistence, i.e. gitlab. Now the content is held in the cache of a context manager.