APPSRE-8094: Test Vector for fluentd replacement
For this test, we will run vector and fluentd side-by-side (each in
their own sidecar). This test will not be for a long period of time -
just long enough to understand if the Vector config (and helm/deployment
config) is working correctly.

Vector maintains the same, high level functionality as fluentd in that
it will still deliver events to Slack, Google Chat and Cloudwatch logs.
In this effort, the message format has been unified between gchat and
slack (per team consensus).

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>