APPSRE-8094: Fix issues from first test
This commit will:

Expose the `POD_NAME` envvar even if it's not a google chat type log
sink (because it's now needed for Vector) as the log messages are
unified. So if neither vector nor google chat are used, the envvar will
still be present. This should not change too much, but did require the
Helm templates to be updated.

Fixes a critical bug in the vector.conf creation (in the init
container): There were an insufficient number of escaped backslashes
that resulted in an invalid config. It must be `\\\\\""`, not `\\\""`.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>