Refactor aws ami cleanup (#4615)
Refactored the ami cleanup integration so that it searches
directly in the AWS accounts to find if an AMI is still being
used in the latest version of a launch template instead of
searching it in app-interface ASG definitions. We're using the
AMIs we build in ASGs defined outside app-interface (Jenkins,
bastion) so we need a new way to make sure that it is safe to
be deleted. In order to know where to look for Launch templates,
we go into the sharing section of the account where we want to
cleanup to get aws accounts in which these AMIs may be used.

This patch also refactors a bit the integration:

* to use typed queries to get aws accounts.
* to minimise the use of the non-typed AWSApi.

APPSRE-10617

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>